requirements.txt
streamlit
faiss-cpu
requests
sentence-transformers
PyPDF2
pdf2image
pillow
pytesseract
